line_iot/requirements.txt
line_bot_sdk
boto3